+++ lib.php
@@ -35,7 +35,7 @@ define('LOCAL_BOOSTNAVIGATION_COLLAPSEICON_YES', 2);
  * @param global_navigation $navigation
  */
 function local_boostnavigation_extend_navigation(global_navigation $navigation) {
-    global $CFG, $PAGE, $COURSE;
+    global $CFG, $PAGE, $COURSE, $USER;
 
     // Fetch config.
     $config = get_config('local_boostnavigation');
@@ -133,50 +133,224 @@ function local_boostnavigation_extend_navigation(global_navigation $navigation) 
                 }
             }
         }
-    }
 
-    // Check if admin wanted us to collapse the mycourses node.
-    // We won't support the setting navshowmycoursecategories in this feature as this would have complicated the feature's
-    // JavaScript code quite heavily.
-    if (isset($config->collapsemycoursesnode) && $config->collapsemycoursesnode == true
+        // Otherwise.
+    } else {
+        // Check if admin wanted us to show all of my courses in Boost's nav drawer regardless of the course progress and then to
+        // apply the block_myoverview filters to the mycourses root node list in Boost's nav drawer.
+        // We won't support the setting navshowmycoursecategories in this feature. It would be possible to support it similar to
+        // $config->removemycoursesnode and as it has been done before in local_boostcoc,
+        // but we want to keep things simple here now.
+        if (isset($config->modifymycoursesrootnodeshowfiltered) && $config->modifymycoursesrootnodeshowfiltered == true
             && $CFG->navshowmycoursecategories == false) {
-        // If yes, do it.
-        if ($mycoursesnode) {
-            // Remember the collapsible node for JavaScript.
-            $collapsenodesforjs[] = 'mycourses';
-            // Add the localboostnavigationcollapsibleparent class to the mycourses node.
-            $mycoursesnode->add_class('localboostnavigationcollapsibleparent');
-            // Get the user preference for the collapse state of the mycourses node and add the localboostnavigationcollapsedparent
-            // and localboostnavigationcollapsedchild classes accordingly.
-            // Additionally, add the localboostnavigationcollapsiblechild class to all child nodes.
-            $mycoursesnode->add_class('localboostnavigationcollapsibleparent');
-            $userprefmycoursesnode = get_user_preferences('local_boostnavigation-collapse_mycoursesnode',
-                    $config->collapsemycoursesnodedefault);
-            if ($userprefmycoursesnode == 1) {
-                $mycoursesnode->add_class('localboostnavigationcollapsedparent');
-                foreach ($mycourseschildrennodeskeys as $k) {
-                    $childnode = $mycoursesnode->get($k);
-                    $childnode->add_class('localboostnavigationcollapsiblechild');
-                    $childnode->add_class('localboostnavigationcollapsedchild');
-                }
-            } else {
-                foreach ($mycourseschildrennodeskeys as $k) {
-                    $childnode = $mycoursesnode->get($k);
-                    $childnode->add_class('localboostnavigationcollapsiblechild');
+            // If yes, do it.
+            if ($mycoursesnode) {
+                // Get list of my courses with all necessary fields.
+                $mycourses = enrol_get_my_courses('enddate');
+                // Do only if I am enrolled in at least one course.
+                if (count($mycourses) > 0) {
+                    // Get the current filter setting, falling back to Moodle's default COURSE_TIMELINE_ALL
+                    // if the user preference does not exist yet.
+                    $currentfilter = get_user_preferences('block_myoverview_user_grouping_preference',
+                            COURSE_TIMELINE_ALL);
+
+                    // Do only if we are dealing with another filter than COURSE_TIMELINE_INPROGRESS as this is already the list
+                    // how it is delivered by Moodle core.
+                    if ($currentfilter != COURSE_TIMELINE_INPROGRESS) {
+                        // If the filter is set to show favourite courses.
+                        if ($currentfilter == COURSE_FAVOURITES) {
+                            // Get the favourite courses.
+                            $favouritecourseids = [];
+                            $ufservice = \core_favourites\service_factory::get_service_for_user_context(
+                                \context_user::instance($USER->id)
+                            );
+                            $favourites = $ufservice->find_favourites_by_type('core_course', 'courses');
+                            if ($favourites) {
+                                $favouritecourseids = array_map(
+                                    function ($favourite) {
+                                        return $favourite->itemid;
+                                    }, $favourites);
+                            }
+                            list($filteredcourses, $processedcount) = course_filter_courses_by_favourites(
+                                    $mycourses,
+                                    $favouritecourseids
+                            );
+
+                            // If the filter is set to show custom course fields.
+                        } else if ($currentfilter == COURSE_CUSTOMFIELD) {
+                            // Get the filtered courses.
+                            $customfieldname = get_config('block_myoverview', 'customfiltergrouping');
+                            $customfieldvalue = get_user_preferences('block_myoverview_user_grouping_customfieldvalue_preference');
+                            list($filteredcourses, $processedcount) = course_filter_courses_by_customfield(
+                                    $mycourses,
+                                    $customfieldname,
+                                    $customfieldvalue
+                            );
+
+                            // If the filter is set to any other value.
+                        } else {
+                            // Get the filtered courses.
+                            list($filteredcourses, $processedcount) = course_filter_courses_by_timeline_classification(
+                                    $mycourses,
+                                    $currentfilter
+                            );
+                        }
+
+                        // If there are any filtered courses to be shown.
+                        if (count($filteredcourses) > 0) {
+                            // Reduce complexity of the list of filtered courses as we really only need the course IDs.
+                            $filteredcoursesids = array_map(
+                                function ($course) {
+                                    return $course->id;
+                                }, $filteredcourses);
+
+                            // Verify that all courses below the mycourses node which should be shown are really shown and
+                            // all courses which should not be shown are not shown.
+                            foreach ($mycourseschildrennodeskeys as $k) {
+                                // If the node is in the list of shown courses.
+                                if (in_array($k, $filteredcoursesids)) {
+                                    // Show the course node.
+                                    $mycoursesnode->get($k)->showinflatnavigation = true;
+
+                                    // Otherwise.
+                                } else {
+                                    // Hide the course node.
+                                    $mycoursesnode->get($k)->showinflatnavigation = false;
+                                }
+                            }
+
+                            // Otherwise.
+                        } else {
+                            // Simply hide all courses.
+                            foreach ($mycourseschildrennodeskeys as $k) {
+                                $mycoursesnode->get($k)->showinflatnavigation = false;
+                            }
+                        }
+                    }
                 }
             }
+        }
 
-            // Check if admin really wanted to show an icon in the parent node and indent the parent node.
-            // Case: LOCAL_BOOSTNAVIGATION_COLLAPSEICON_YES) - Icon and indent is already fine.
-            // Case: LOCAL_BOOSTNAVIGATION_COLLAPSEICON_JUSTINDENT - Icon has to be removed, but indent is fine.
-            // Note that the icon is removed by setting it to i/navigationitem which is mapped it fa-fw
-            // and which is the same as the navigation_node constructor sets if the icon is set to null.
-            if ($config->collapsemycoursesnodeicon == LOCAL_BOOSTNAVIGATION_COLLAPSEICON_JUSTINDENT) {
-                $mycoursesnode->icon = new pix_icon('i/navigationitem', '');
-                // Case: LOCAL_BOOSTNAVIGATION_COLLAPSEICON_NONE - Icon and indent have to be removed.
-            } else if ($config->collapsemycoursesnodeicon == LOCAL_BOOSTNAVIGATION_COLLAPSEICON_NONE) {
-                $mycoursesnode->icon = new pix_icon('i/navigationitem', '');
-                $mycoursesnode->add_class('localboostnavigationcollapsibleparentforcenoindent');
+        // Check if admin wanted us to add the active filters hint root node or to add the change filter link root node
+        // to the mycourses list in Boost's nav drawer.
+        if ((isset($config->modifymycoursesrootnodefilterhint) && $config->modifymycoursesrootnodefilterhint == true) ||
+            (isset($config->modifymycoursesrootnodeshowfiltered) && $config->modifymycoursesrootnodeshowfiltered == true &&
+                isset($config->modifymycoursesrootnodefilterlink) && $config->modifymycoursesrootnodefilterlink == true)) {
+            // If yes, do it.
+            if ($mycoursesnode) {
+                // Do only if I am enrolled in at least one course.
+                if (count($mycourseschildrennodeskeys) > 0) {
+                    // Prepare string.
+                    $mycourseshintstring = '';
+
+                    // If enabled, create active filters hint.
+                    if (isset($config->modifymycoursesrootnodefilterhint) && $config->modifymycoursesrootnodefilterhint == true) {
+                        // If showing filtered courses is enabled, build active filters hint string.
+                        if (isset($config->modifymycoursesrootnodeshowfiltered) &&
+                                $config->modifymycoursesrootnodeshowfiltered == true) {
+                            // Get the current filter setting, falling back to Moodle's default COURSE_TIMELINE_ALL
+                            // if the user preference does not exist yet.
+                            $currentfilter = get_user_preferences('block_myoverview_user_grouping_preference',
+                                    COURSE_TIMELINE_ALL);
+
+                            // Build the active filters hint.
+                            $mycourseshintstring = get_string('mycoursesrootnodefilterhintenabledcurrentfilter',
+                                    'local_boostnavigation');
+                            $mycourseshintstring .= ' ';
+                            $mycourseshintstring .= get_string($currentfilter, 'block_myoverview');
+
+                            // Otherwise hardcode the inprogress filter.
+                        } else {
+                            // Build the active filters hint.
+                            $mycourseshintstring = get_string('mycoursesrootnodefilterhintenabledcourselist',
+                                    'local_boostnavigation');
+                            $mycourseshintstring .= ' ';
+                            $mycourseshintstring .= get_string(COURSE_TIMELINE_INPROGRESS, 'block_myoverview');
+                        }
+                    }
+
+                    // If enabled and not on Dashboard, create change filters link.
+                    if (isset($config->modifymycoursesrootnodeshowfiltered) &&
+                            $config->modifymycoursesrootnodeshowfiltered == true &&
+                            isset($config->modifymycoursesrootnodefilterlink) &&
+                            $config->modifymycoursesrootnodefilterlink == true &&
+                            $PAGE->pagelayout !== 'mydashboard') {
+
+                        // Add line break if the hint setting is enabled as well.
+                        if (isset($config->modifymycoursesrootnodefilterhint) &&
+                                $config->modifymycoursesrootnodefilterhint == true) {
+                            $mycourseshintstring .= html_writer::empty_tag('br');
+                        }
+
+                        $url = new moodle_url('/my/'); // Link target: Dashboard.
+                        $mycourseshintstring .= html_writer::link($url, get_string('mycoursesrootnodefilterlink',
+                                'local_boostnavigation'));
+                    }
+
+                    // If we really have a node to recreate now.
+                    if ($mycourseshintstring != '') {
+                        // Create new navigation node.
+                        // (use TYPE_COURSE to get the correct indent instead of TYPE_CUSTOM which would be semantically correct).
+                        $mycourseshintnode = navigation_node::create($mycourseshintstring, null, global_navigation::TYPE_COURSE, null,
+                            'localboostnavigationactivefiltershint', new pix_icon('i/filter', ''));
+
+                        // Show the navigation node in Boost's nav drawer.
+                        $mycourseshintnode->showinflatnavigation = true;
+
+                        // Add the node to the mycourses list in Boost's nav drawer
+                        // (will be added at the end where we want it to be).
+                        $mycoursesnode->add_node($mycourseshintnode);
+                    }
+                }
+            }
+        }
+
+        // Check if admin wanted us to collapse the mycourses node.
+        // We won't support the setting navshowmycoursecategories in this feature as this would have complicated the feature's
+        // JavaScript code quite heavily.
+        if (isset($config->collapsemycoursesnode) && $config->collapsemycoursesnode == true
+            && $CFG->navshowmycoursecategories == false) {
+            // If yes, do it.
+            if ($mycoursesnode) {
+                // Remember the collapsible node for JavaScript.
+                $collapsenodesforjs[] = 'mycourses';
+                // Add the localboostnavigationcollapsibleparent class to the mycourses node.
+                $mycoursesnode->add_class('localboostnavigationcollapsibleparent');
+                // Get the user preference for the collapse state of the mycourses node and add the
+                // localboostnavigationcollapsedparent and localboostnavigationcollapsedchild classes accordingly.
+                // Additionally, add the localboostnavigationcollapsiblechild class to all child nodes.
+                $mycoursesnode->add_class('localboostnavigationcollapsibleparent');
+                $userprefmycoursesnode = get_user_preferences('local_boostnavigation-collapse_mycoursesnode',
+                    $config->collapsemycoursesnodedefault);
+                if ($userprefmycoursesnode == 1) {
+                    $mycoursesnode->add_class('localboostnavigationcollapsedparent');
+                    foreach ($mycourseschildrennodeskeys as $k) {
+                        $childnode = $mycoursesnode->get($k);
+                        $childnode->add_class('localboostnavigationcollapsiblechild');
+                        $childnode->add_class('localboostnavigationcollapsedchild');
+                    }
+                } else {
+                    foreach ($mycourseschildrennodeskeys as $k) {
+                        $childnode = $mycoursesnode->get($k);
+                        $childnode->add_class('localboostnavigationcollapsiblechild');
+                    }
+                }
+                if (isset($mycourseshintnode)) {
+                    $mycourseshintnode->add_class('localboostnavigationcollapsiblechild');
+                }
+
+                // Check if admin really wanted to show an icon in the parent node and indent the parent node.
+                // Case: LOCAL_BOOSTNAVIGATION_COLLAPSEICON_YES) - Icon and indent is already fine.
+                // Case: LOCAL_BOOSTNAVIGATION_COLLAPSEICON_JUSTINDENT - Icon has to be removed, but indent is fine.
+                // Note that the icon is removed by setting it to i/navigationitem which is mapped it fa-fw
+                // and which is the same as the navigation_node constructor sets if the icon is set to null.
+                if ($config->collapsemycoursesnodeicon == LOCAL_BOOSTNAVIGATION_COLLAPSEICON_JUSTINDENT) {
+                    $mycoursesnode->icon = new pix_icon('i/navigationitem', '');
+                    // Case: LOCAL_BOOSTNAVIGATION_COLLAPSEICON_NONE - Icon and indent have to be removed.
+                } else if ($config->collapsemycoursesnodeicon == LOCAL_BOOSTNAVIGATION_COLLAPSEICON_NONE) {
+                    $mycoursesnode->icon = new pix_icon('i/navigationitem', '');
+                    $mycoursesnode->add_class('localboostnavigationcollapsibleparentforcenoindent');
+                }
             }
         }
     }
@@ -864,4 +1038,23 @@ function local_boostnavigation_reset_fontawesome_icon_map() {
     $cache->delete($mapkey);
     // And rebuild it brutally.
     $instance->get_icon_name_map();
+}
+
+/**
+ * Hack into the Moodle page build process as early as possible to modify $CFG by leveraging Moodle's *_before_http_headers() hook.
+ */
+function local_boostnavigation_before_http_headers() {
+    global $CFG;
+
+    // Fetch config.
+    $config = get_config('local_boostnavigation');
+
+    // If the plugin setting modifymycoursesrootnodeshowfiltered (which is explained is enabled above in
+    // local_boostnavigation_extend_navigation()), then it is necessary to avoid that the nav drawer course list length is limited.
+    // We realize that by setting $CFG->navcourselimit to a very high value and we do this here and automatically so that the
+    // admin can't forget it.
+    if (isset($config->modifymycoursesrootnodeshowfiltered) && $config->modifymycoursesrootnodeshowfiltered == true
+            && $CFG->navshowmycoursecategories == false) {
+        $CFG->navcourselimit = 100000;
+    }
 }